requirements.txt
python-slugify
BeautifulSoup